Changes for Firebase and Fewer phonemes
In schedule, duplicated already existing panel to contain fewer phonemes. User value now affects the state management (added in helpers). This is done in order to track the user value in when using the activity logger for Firebase. This logger is used to track data from the event listeners in Step and Test.